branch code style refactor
