+++ kubehub/serializers/vm_group_serializer.py
@@ -1,15 +1,15 @@
 from rest_framework import serializers
-from ..models.VmGroup import VmGroups
+from ..models.vm_group import VMGroup
 
 
-class VmGroupsSerializer(serializers.ModelSerializer):
+class VMGroupSerializer(serializers.ModelSerializer):
     class Meta:
-        model = VmGroups
+        model = VMGroup
         fields = ('id', 'vm_group_name', 'vm_ip')
 
     def create(self, validated_data):
-        vm_groups = VmGroups.objects.create(**validated_data)
-        return vm_groups
+        vm_group = VMGroup.objects.create(**validated_data)
+        return vm_group
 
     def update(self, instance, validated_data):
         instance.vm_group_name = validated_data.get('vm_group_name', instance.vm_group_name)
